% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haikunator.R
\name{haikunate}
\alias{haikunate}
\title{Generate Heroku-like Memorable Random Names}
\usage{
haikunate(token_length = 4, delimiter = "-", seed_value = NA)
}
\arguments{
\item{token_length}{Token Length}

\item{delimiter}{Delimiter}

\item{seed_value}{Seed Value}
}
\examples{
haikunate(5)
haikunate(3,'.',123)
}
